# ARROW-6798: [CI] [Rust] Build image contains cached dependencies