This PR adds `sccache` and builds many of the project dependencies in the base image to avoid downloading and compiling them each time the project builds in CI.