@kou 2019-10-09
I'm testing this.
Could you show how to confirm whether cache is used or not?