@andygrove 2019-10-09
Thanks @kou  ... my advice would be to run `docker-compose build rust` then `docker-compose run rust` before and after this PR and count how many crates get *downloaded* (not compiled). You should see a significant reduction after this PR.