@andygrove 2019-10-13
@kou Are you stating timing just for "run" or for "build" and "run" ? Overall this will take longer, but the "run" part should have been faster. This is only useful if the build step gets cached.